+++ yellowstone-grpc-geyser/src/grpc.rs
@@ -1,23 +1,17 @@
 use {
     crate::{
         config::{ConfigBlockFailAction, ConfigGrpc, ConfigGrpcFilters},
-        filters::{Filter, FilterAccountsDataSlice},
+        filters::Filter,
+        message::{Message, MessageBlockMeta, MessageEntry, MessageSlot, MessageTransactionInfo},
         metrics::{self, DebugClientMessage},
         version::GrpcVersionInfo,
-    },
-    agave_geyser_plugin_interface::geyser_plugin_interface::{
-        ReplicaAccountInfoV3, ReplicaBlockInfoV4, ReplicaEntryInfoV2, ReplicaTransactionInfoV2,
-        SlotStatus,
     },
     anyhow::Context,
     log::{error, info},
     solana_sdk::{
-        clock::{Slot, UnixTimestamp, MAX_RECENT_BLOCKHASHES},
+        clock::{Slot, MAX_RECENT_BLOCKHASHES},
         pubkey::Pubkey,
-        signature::Signature,
-        transaction::SanitizedTransaction,
     },
-    solana_transaction_status::{Reward, TransactionStatusMeta},
     std::{
         collections::{BTreeMap, HashMap},
         sync::{
@@ -42,456 +36,15 @@ use {
         Request, Response, Result as TonicResult, Status, Streaming,
     },
     tonic_health::server::health_reporter,
-    yellowstone_grpc_proto::{
-        convert_to,
-        prelude::{
-            geyser_server::{Geyser, GeyserServer},
-            subscribe_update::UpdateOneof,
-            CommitmentLevel, GetBlockHeightRequest, GetBlockHeightResponse,
-            GetLatestBlockhashRequest, GetLatestBlockhashResponse, GetSlotRequest, GetSlotResponse,
-            GetVersionRequest, GetVersionResponse, IsBlockhashValidRequest,
-            IsBlockhashValidResponse, PingRequest, PongResponse, SubscribeRequest, SubscribeUpdate,
-            SubscribeUpdateAccount, SubscribeUpdateAccountInfo, SubscribeUpdateBlock,
-            SubscribeUpdateBlockMeta, SubscribeUpdateEntry, SubscribeUpdatePing,
-            SubscribeUpdateSlot, SubscribeUpdateTransaction, SubscribeUpdateTransactionInfo,
-            SubscribeUpdateTransactionStatus, TransactionError as SubscribeUpdateTransactionError,
-        },
+    yellowstone_grpc_proto::prelude::{
+        geyser_server::{Geyser, GeyserServer},
+        subscribe_update::UpdateOneof,
+        CommitmentLevel, GetBlockHeightRequest, GetBlockHeightResponse, GetLatestBlockhashRequest,
+        GetLatestBlockhashResponse, GetSlotRequest, GetSlotResponse, GetVersionRequest,
+        GetVersionResponse, IsBlockhashValidRequest, IsBlockhashValidResponse, PingRequest,
+        PongResponse, SubscribeRequest, SubscribeUpdate, SubscribeUpdatePing,
     },
 };
-
-#[derive(Debug, Clone)]
-pub struct MessageAccountInfo {
-    pub pubkey: Pubkey,
-    pub lamports: u64,
-    pub owner: Pubkey,
-    pub executable: bool,
-    pub rent_epoch: u64,
-    pub data: Vec<u8>,
-    pub write_version: u64,
-    pub txn_signature: Option<Signature>,
-}
-
-impl MessageAccountInfo {
-    fn to_proto(
-        &self,
-        accounts_data_slice: &[FilterAccountsDataSlice],
-    ) -> SubscribeUpdateAccountInfo {
-        let data = if accounts_data_slice.is_empty() {
-            self.data.clone()
-        } else {
-            let mut data = Vec::with_capacity(accounts_data_slice.iter().map(|ds| ds.length).sum());
-            for data_slice in accounts_data_slice {
-                if self.data.len() >= data_slice.end {
-                    data.extend_from_slice(&self.data[data_slice.start..data_slice.end]);
-                }
-            }
-            data
-        };
-        SubscribeUpdateAccountInfo {
-            pubkey: self.pubkey.as_ref().into(),
-            lamports: self.lamports,
-            owner: self.owner.as_ref().into(),
-            executable: self.executable,
-            rent_epoch: self.rent_epoch,
-            data,
-            write_version: self.write_version,
-            txn_signature: self.txn_signature.map(|s| s.as_ref().into()),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct MessageAccount {
-    pub account: MessageAccountInfo,
-    pub slot: u64,
-    pub is_startup: bool,
-}
-
-impl<'a> From<(&'a ReplicaAccountInfoV3<'a>, u64, bool)> for MessageAccount {
-    fn from((account, slot, is_startup): (&'a ReplicaAccountInfoV3<'a>, u64, bool)) -> Self {
-        Self {
-            account: MessageAccountInfo {
-                pubkey: Pubkey::try_from(account.pubkey).expect("valid Pubkey"),
-                lamports: account.lamports,
-                owner: Pubkey::try_from(account.owner).expect("valid Pubkey"),
-                executable: account.executable,
-                rent_epoch: account.rent_epoch,
-                data: account.data.into(),
-                write_version: account.write_version,
-                txn_signature: account.txn.map(|txn| *txn.signature()),
-            },
-            slot,
-            is_startup,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct MessageSlot {
-    pub slot: u64,
-    pub parent: Option<u64>,
-    pub status: CommitmentLevel,
-}
-
-impl From<(u64, Option<u64>, SlotStatus)> for MessageSlot {
-    fn from((slot, parent, status): (u64, Option<u64>, SlotStatus)) -> Self {
-        Self {
-            slot,
-            parent,
-            status: match status {
-                SlotStatus::Processed => CommitmentLevel::Processed,
-                SlotStatus::Confirmed => CommitmentLevel::Confirmed,
-                SlotStatus::Rooted => CommitmentLevel::Finalized,
-            },
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct MessageTransactionInfo {
-    pub signature: Signature,
-    pub is_vote: bool,
-    pub transaction: SanitizedTransaction,
-    pub meta: TransactionStatusMeta,
-    pub index: usize,
-}
-
-impl MessageTransactionInfo {
-    fn to_proto(&self) -> SubscribeUpdateTransactionInfo {
-        SubscribeUpdateTransactionInfo {
-            signature: self.signature.as_ref().into(),
-            is_vote: self.is_vote,
-            transaction: Some(convert_to::create_transaction(&self.transaction)),
-            meta: Some(convert_to::create_transaction_meta(&self.meta)),
-            index: self.index as u64,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct MessageTransaction {
-    pub transaction: MessageTransactionInfo,
-    pub slot: u64,
-}
-
-impl<'a> From<(&'a ReplicaTransactionInfoV2<'a>, u64)> for MessageTransaction {
-    fn from((transaction, slot): (&'a ReplicaTransactionInfoV2<'a>, u64)) -> Self {
-        Self {
-            transaction: MessageTransactionInfo {
-                signature: *transaction.signature,
-                is_vote: transaction.is_vote,
-                transaction: transaction.transaction.clone(),
-                meta: transaction.transaction_status_meta.clone(),
-                index: transaction.index,
-            },
-            slot,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct MessageEntry {
-    pub slot: u64,
-    pub index: usize,
-    pub num_hashes: u64,
-    pub hash: Vec<u8>,
-    pub executed_transaction_count: u64,
-    pub starting_transaction_index: u64,
-}
-
-impl From<&ReplicaEntryInfoV2<'_>> for MessageEntry {
-    fn from(entry: &ReplicaEntryInfoV2) -> Self {
-        Self {
-            slot: entry.slot,
-            index: entry.index,
-            num_hashes: entry.num_hashes,
-            hash: entry.hash.into(),
-            executed_transaction_count: entry.executed_transaction_count,
-            starting_transaction_index: entry
-                .starting_transaction_index
-                .try_into()
-                .expect("failed convert usize to u64"),
-        }
-    }
-}
-
-impl MessageEntry {
-    fn to_proto(&self) -> SubscribeUpdateEntry {
-        SubscribeUpdateEntry {
-            slot: self.slot,
-            index: self.index as u64,
-            num_hashes: self.num_hashes,
-            hash: self.hash.clone(),
-            executed_transaction_count: self.executed_transaction_count,
-            starting_transaction_index: self.starting_transaction_index,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct MessageBlock {
-    pub parent_slot: u64,
-    pub slot: u64,
-    pub parent_blockhash: String,
-    pub blockhash: String,
-    pub rewards: Vec<Reward>,
-    pub num_partitions: Option<u64>,
-    pub block_time: Option<UnixTimestamp>,
-    pub block_height: Option<u64>,
-    pub executed_transaction_count: u64,
-    pub transactions: Vec<MessageTransactionInfo>,
-    pub updated_account_count: u64,
-    pub accounts: Vec<MessageAccountInfo>,
-    pub entries_count: u64,
-    pub entries: Vec<MessageEntry>,
-}
-
-impl
-    From<(
-        MessageBlockMeta,
-        Vec<MessageTransactionInfo>,
-        Vec<MessageAccountInfo>,
-        Vec<MessageEntry>,
-    )> for MessageBlock
-{
-    fn from(
-        (blockinfo, transactions, accounts, entries): (
-            MessageBlockMeta,
-            Vec<MessageTransactionInfo>,
-            Vec<MessageAccountInfo>,
-            Vec<MessageEntry>,
-        ),
-    ) -> Self {
-        Self {
-            parent_slot: blockinfo.parent_slot,
-            slot: blockinfo.slot,
-            blockhash: blockinfo.blockhash,
-            parent_blockhash: blockinfo.parent_blockhash,
-            rewards: blockinfo.rewards,
-            num_partitions: blockinfo.num_partitions,
-            block_time: blockinfo.block_time,
-            block_height: blockinfo.block_height,
-            executed_transaction_count: blockinfo.executed_transaction_count,
-            transactions,
-            updated_account_count: accounts.len() as u64,
-            accounts,
-            entries_count: entries.len() as u64,
-            entries,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct MessageBlockMeta {
-    pub parent_slot: u64,
-    pub slot: u64,
-    pub parent_blockhash: String,
-    pub blockhash: String,
-    pub rewards: Vec<Reward>,
-    pub num_partitions: Option<u64>,
-    pub block_time: Option<UnixTimestamp>,
-    pub block_height: Option<u64>,
-    pub executed_transaction_count: u64,
-    pub entries_count: u64,
-}
-
-impl<'a> From<&'a ReplicaBlockInfoV4<'a>> for MessageBlockMeta {
-    fn from(blockinfo: &'a ReplicaBlockInfoV4<'a>) -> Self {
-        Self {
-            parent_slot: blockinfo.parent_slot,
-            slot: blockinfo.slot,
-            parent_blockhash: blockinfo.parent_blockhash.to_string(),
-            blockhash: blockinfo.blockhash.to_string(),
-            rewards: blockinfo.rewards.rewards.clone(),
-            num_partitions: blockinfo.rewards.num_partitions,
-            block_time: blockinfo.block_time,
-            block_height: blockinfo.block_height,
-            executed_transaction_count: blockinfo.executed_transaction_count,
-            entries_count: blockinfo.entry_count,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-#[allow(clippy::large_enum_variant)]
-pub enum Message {
-    Slot(MessageSlot),
-    Account(MessageAccount),
-    Transaction(MessageTransaction),
-    Entry(MessageEntry),
-    Block(MessageBlock),
-    BlockMeta(MessageBlockMeta),
-}
-
-impl Message {
-    pub const fn get_slot(&self) -> u64 {
-        match self {
-            Self::Slot(msg) => msg.slot,
-            Self::Account(msg) => msg.slot,
-            Self::Transaction(msg) => msg.slot,
-            Self::Entry(msg) => msg.slot,
-            Self::Block(msg) => msg.slot,
-            Self::BlockMeta(msg) => msg.slot,
-        }
-    }
-
-    pub const fn kind(&self) -> &'static str {
-        match self {
-            Self::Slot(_) => "Slot",
-            Self::Account(_) => "Account",
-            Self::Transaction(_) => "Transaction",
-            Self::Entry(_) => "Entry",
-            Self::Block(_) => "Block",
-            Self::BlockMeta(_) => "BlockMeta",
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct MessageBlockRef<'a> {
-    pub parent_slot: u64,
-    pub slot: u64,
-    pub parent_blockhash: &'a String,
-    pub blockhash: &'a String,
-    pub rewards: &'a Vec<Reward>,
-    pub num_partitions: Option<u64>,
-    pub block_time: Option<UnixTimestamp>,
-    pub block_height: Option<u64>,
-    pub executed_transaction_count: u64,
-    pub transactions: Vec<&'a MessageTransactionInfo>,
-    pub updated_account_count: u64,
-    pub accounts: Vec<&'a MessageAccountInfo>,
-    pub entries_count: u64,
-    pub entries: Vec<&'a MessageEntry>,
-}
-
-impl<'a>
-    From<(
-        &'a MessageBlock,
-        Vec<&'a MessageTransactionInfo>,
-        Vec<&'a MessageAccountInfo>,
-        Vec<&'a MessageEntry>,
-    )> for MessageBlockRef<'a>
-{
-    fn from(
-        (block, transactions, accounts, entries): (
-            &'a MessageBlock,
-            Vec<&'a MessageTransactionInfo>,
-            Vec<&'a MessageAccountInfo>,
-            Vec<&'a MessageEntry>,
-        ),
-    ) -> Self {
-        Self {
-            parent_slot: block.parent_slot,
-            slot: block.slot,
-            parent_blockhash: &block.parent_blockhash,
-            blockhash: &block.blockhash,
-            rewards: &block.rewards,
-            num_partitions: block.num_partitions,
-            block_time: block.block_time,
-            block_height: block.block_height,
-            executed_transaction_count: block.executed_transaction_count,
-            transactions,
-            updated_account_count: block.updated_account_count,
-            accounts,
-            entries_count: block.entries_count,
-            entries,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-#[allow(clippy::large_enum_variant)]
-pub enum MessageRef<'a> {
-    Slot(&'a MessageSlot),
-    Account(&'a MessageAccount),
-    Transaction(&'a MessageTransaction),
-    TransactionStatus(&'a MessageTransaction),
-    Entry(&'a MessageEntry),
-    Block(MessageBlockRef<'a>),
-    BlockMeta(&'a MessageBlockMeta),
-}
-
-impl<'a> MessageRef<'a> {
-    pub fn to_proto(&self, accounts_data_slice: &[FilterAccountsDataSlice]) -> UpdateOneof {
-        match self {
-            Self::Slot(message) => UpdateOneof::Slot(SubscribeUpdateSlot {
-                slot: message.slot,
-                parent: message.parent,
-                status: message.status as i32,
-            }),
-            Self::Account(message) => UpdateOneof::Account(SubscribeUpdateAccount {
-                account: Some(message.account.to_proto(accounts_data_slice)),
-                slot: message.slot,
-                is_startup: message.is_startup,
-            }),
-            Self::Transaction(message) => UpdateOneof::Transaction(SubscribeUpdateTransaction {
-                transaction: Some(message.transaction.to_proto()),
-                slot: message.slot,
-            }),
-            Self::TransactionStatus(message) => {
-                UpdateOneof::TransactionStatus(SubscribeUpdateTransactionStatus {
-                    slot: message.slot,
-                    signature: message.transaction.signature.as_ref().into(),
-                    is_vote: message.transaction.is_vote,
-                    index: message.transaction.index as u64,
-                    err: match &message.transaction.meta.status {
-                        Ok(()) => None,
-                        Err(err) => Some(SubscribeUpdateTransactionError {
-                            err: bincode::serialize(&err)
-                                .expect("transaction error to serialize to bytes"),
-                        }),
-                    },
-                })
-            }
-            Self::Entry(message) => UpdateOneof::Entry(message.to_proto()),
-            Self::Block(message) => UpdateOneof::Block(SubscribeUpdateBlock {
-                slot: message.slot,
-                blockhash: message.blockhash.clone(),
-                rewards: Some(convert_to::create_rewards_obj(
-                    message.rewards.as_slice(),
-                    message.num_partitions,
-                )),
-                block_time: message.block_time.map(convert_to::create_timestamp),
-                block_height: message.block_height.map(convert_to::create_block_height),
-                parent_slot: message.parent_slot,
-                parent_blockhash: message.parent_blockhash.clone(),
-                executed_transaction_count: message.executed_transaction_count,
-                transactions: message
-                    .transactions
-                    .iter()
-                    .map(|tx| tx.to_proto())
-                    .collect(),
-                updated_account_count: message.updated_account_count,
-                accounts: message
-                    .accounts
-                    .iter()
-                    .map(|acc| acc.to_proto(accounts_data_slice))
-                    .collect(),
-                entries_count: message.entries_count,
-                entries: message
-                    .entries
-                    .iter()
-                    .map(|entry| entry.to_proto())
-                    .collect(),
-            }),
-            Self::BlockMeta(message) => UpdateOneof::BlockMeta(SubscribeUpdateBlockMeta {
-                slot: message.slot,
-                blockhash: message.blockhash.clone(),
-                rewards: Some(convert_to::create_rewards_obj(
-                    message.rewards.as_slice(),
-                    message.num_partitions,
-                )),
-                block_time: message.block_time.map(convert_to::create_timestamp),
-                block_height: message.block_height.map(convert_to::create_block_height),
-                parent_slot: message.parent_slot,
-                parent_blockhash: message.parent_blockhash.clone(),
-                executed_transaction_count: message.executed_transaction_count,
-                entries_count: message.entries_count,
-            }),
-        }
-    }
-}
 
 #[derive(Debug)]
 struct BlockhashStatus {
@@ -514,7 +67,7 @@ impl BlockhashStatus {
 
 #[derive(Debug, Default)]
 struct BlockMetaStorageInner {
-    blocks: HashMap<u64, MessageBlockMeta>,
+    blocks: HashMap<u64, Arc<MessageBlockMeta>>,
     blockhashes: HashMap<String, BlockhashStatus>,
     processed: Option<u64>,
     confirmed: Option<u64>,
@@ -669,11 +222,11 @@ impl BlockMetaStorage {
 
 #[derive(Debug, Default)]
 struct SlotMessages {
-    messages: Vec<Option<Arc<Message>>>, // Option is used for accounts with low write_version
-    block_meta: Option<MessageBlockMeta>,
-    transactions: Vec<MessageTransactionInfo>,
+    messages: Vec<Option<Message>>, // Option is used for accounts with low write_version
+    block_meta: Option<Arc<MessageBlockMeta>>,
+    transactions: Vec<Arc<MessageTransactionInfo>>,
     accounts_dedup: HashMap<Pubkey, (u64, usize)>, // (write_version, message_index)
-    entries: Vec<MessageEntry>,
+    entries: Vec<Arc<MessageEntry>>,
     sealed: bool,
     entries_count: usize,
     confirmed_at: Option<usize>,
@@ -684,7 +237,7 @@ struct SlotMessages {
 }
 
 impl SlotMessages {
-    pub fn try_seal(&mut self) -> Option<Arc<Message>> {
+    pub fn try_seal(&mut self) -> Option<Message> {
         if !self.sealed {
             if let Some(block_meta) = &self.block_meta {
                 let executed_transaction_count = block_meta.executed_transaction_count as usize;
@@ -703,15 +256,15 @@ impl SlotMessages {
 
                     let mut accounts = Vec::with_capacity(self.messages.len());
                     for item in self.messages.iter().flatten() {
-                        if let Message::Account(account) = item.as_ref() {
-                            accounts.push(account.account.clone());
+                        if let Message::Account(account) = item {
+                            accounts.push(Arc::clone(&account.account));
                         }
                     }
 
-                    let message = Arc::new(Message::Block(
-                        (block_meta.clone(), transactions, accounts, entries).into(),
+                    let message = Message::Block(Arc::new(
+                        (Arc::clone(block_meta), transactions, accounts, entries).into(),
                     ));
-                    self.messages.push(Some(Arc::clone(&message)));
+                    self.messages.push(Some(message.clone()));
 
                     self.sealed = true;
                     self.entries_count = entries_count;
@@ -732,7 +285,7 @@ pub struct GrpcService {
     blocks_meta: Option<BlockMetaStorage>,
     subscribe_id: AtomicUsize,
     snapshot_rx: Mutex<Option<crossbeam_channel::Receiver<Box<Message>>>>,
-    broadcast_tx: broadcast::Sender<(CommitmentLevel, Arc<Vec<Arc<Message>>>)>,
+    broadcast_tx: broadcast::Sender<(CommitmentLevel, Arc<Vec<Message>>)>,
     debug_clients_tx: Option<mpsc::UnboundedSender<DebugClientMessage>>,
 }
 
@@ -745,7 +298,7 @@ impl GrpcService {
         is_reload: bool,
     ) -> anyhow::Result<(
         Option<crossbeam_channel::Sender<Box<Message>>>,
-        mpsc::UnboundedSender<Arc<Message>>,
+        mpsc::UnboundedSender<Message>,
         Arc<Notify>,
     )> {
         // Bind service address
@@ -857,9 +410,9 @@ impl GrpcService {
     }
 
     async fn geyser_loop(
-        mut messages_rx: mpsc::UnboundedReceiver<Arc<Message>>,
+        mut messages_rx: mpsc::UnboundedReceiver<Message>,
         blocks_meta_tx: Option<mpsc::UnboundedSender<Message>>,
-        broadcast_tx: broadcast::Sender<(CommitmentLevel, Arc<Vec<Arc<Message>>>)>,
+        broadcast_tx: broadcast::Sender<(CommitmentLevel, Arc<Vec<Message>>)>,
         block_fail_action: ConfigBlockFailAction,
     ) {
         const PROCESSED_MESSAGES_MAX: usize = 31;
@@ -877,19 +430,19 @@ impl GrpcService {
                     metrics::message_queue_size_dec();
 
                     // Update metrics
-                    if let Message::Slot(slot_message) = message.as_ref() {
+                    if let Message::Slot(slot_message) = &message {
                         metrics::update_slot_plugin_status(slot_message.status, slot_message.slot);
                     }
 
                     // Update blocks info
                     if let Some(blocks_meta_tx) = &blocks_meta_tx {
-                        if matches!(message.as_ref(), Message::Slot(_) | Message::BlockMeta(_)) {
-                            let _ = blocks_meta_tx.send(message.as_ref().clone());
+                        if matches!(&message, Message::Slot(_) | Message::BlockMeta(_)) {
+                            let _ = blocks_meta_tx.send(message.clone());
                         }
                     }
 
                     // Remove outdated block reconstruction info
-                    match message.as_ref() {
+                    match &message {
                         // On startup we can receive few Confirmed/Finalized slots without BlockMeta message
                         // With saved first Processed slot we can ignore errors caused by startup process
                         Message::Slot(msg) if processed_first_slot.is_none() && msg.status == CommitmentLevel::Processed => {
@@ -950,7 +503,7 @@ impl GrpcService {
 
                     // Update block reconstruction info
                     let slot_messages = messages.entry(message.get_slot()).or_default();
-                    if let Message::Slot(msg) = message.as_ref() {
+                    if let Message::Slot(msg) = &message {
                         match msg.status {
                             CommitmentLevel::Processed => {
                                 slot_messages.parent_slot = msg.parent;
@@ -963,11 +516,11 @@ impl GrpcService {
                             },
                         }
                     }
-                    if !matches!(message.as_ref(), Message::Slot(_)) {
-                        slot_messages.messages.push(Some(Arc::clone(&message)));
+                    if !matches!(&message, Message::Slot(_)) {
+                        slot_messages.messages.push(Some(message.clone()));
 
                         // If we already build Block message, new message will be a problem
-                        if slot_messages.sealed && !(matches!(message.as_ref(), Message::Entry(_)) && slot_messages.entries_count == 0) {
+                        if slot_messages.sealed && !(matches!(&message, Message::Entry(_)) && slot_messages.entries_count == 0) {
                             metrics::update_invalid_blocks(format!("unexpected message {}", message.kind()));
                             match block_fail_action {
                                 ConfigBlockFailAction::Log => {
@@ -980,7 +533,7 @@ impl GrpcService {
                         }
                     }
                     let mut sealed_block_msg = None;
-                    match message.as_ref() {
+                    match &message {
                         Message::BlockMeta(msg) => {
                             if slot_messages.block_meta.is_some() {
                                 metrics::update_invalid_blocks("unexpected message: BlockMeta (duplicate)");
@@ -993,11 +546,11 @@ impl GrpcService {
                                     }
                                 }
                             }
-                            slot_messages.block_meta = Some(msg.clone());
+                            slot_messages.block_meta = Some(Arc::clone(msg));
                             sealed_block_msg = slot_messages.try_seal();
                         }
                         Message::Transaction(msg) => {
-                            slot_messages.transactions.push(msg.transaction.clone());
+                            slot_messages.transactions.push(Arc::clone(&msg.transaction));
                             sealed_block_msg = slot_messages.try_seal();
                         }
                         // Dedup accounts by max write_version
@@ -1015,7 +568,7 @@ impl GrpcService {
                             }
                         }
                         Message::Entry(msg) => {
-                            slot_messages.entries.push(msg.clone());
+                            slot_messages.entries.push(Arc::clone(msg));
                             sealed_block_msg = slot_messages.try_seal();
                         }
                         _ => {}
@@ -1026,7 +579,7 @@ impl GrpcService {
                     if let Some(sealed_block_msg) = sealed_block_msg {
                         messages_vec.push(sealed_block_msg);
                     }
-                    let slot_status = if let Message::Slot(msg) = message.as_ref() {
+                    let slot_status = if let Message::Slot(msg) = &message {
                         Some((msg.slot, msg.status))
                     } else {
                         None
@@ -1052,18 +605,18 @@ impl GrpcService {
                                 }
 
                                 slots.push(parent);
-                                messages_vec.push(Arc::new(Message::Slot(MessageSlot {
+                                messages_vec.push(Message::Slot(MessageSlot {
                                     slot: parent,
                                     parent: entry.parent_slot,
                                     status,
-                                })));
+                                }));
                                 metrics::missed_status_message_inc(status);
                             }
                         }
                     }
 
                     for message in messages_vec.into_iter().rev() {
-                        if let Message::Slot(slot) = message.as_ref() {
+                        if let Message::Slot(slot) = &message {
                             let (mut confirmed_messages, mut finalized_messages) = match slot.status {
                                 CommitmentLevel::Processed => {
                                     (Vec::with_capacity(1), Vec::with_capacity(1))
@@ -1097,7 +650,7 @@ impl GrpcService {
                             };
 
                             // processed
-                            processed_messages.push(Arc::clone(&message));
+                            processed_messages.push(message.clone());
                             let _ =
                                 broadcast_tx.send((CommitmentLevel::Processed, processed_messages.into()));
                             processed_messages = Vec::with_capacity(PROCESSED_MESSAGES_MAX);
@@ -1106,7 +659,7 @@ impl GrpcService {
                                 .reset(Instant::now() + PROCESSED_MESSAGES_SLEEP);
 
                             // confirmed
-                            confirmed_messages.push(Arc::clone(&message));
+                            confirmed_messages.push(message.clone());
                             let _ =
                                 broadcast_tx.send((CommitmentLevel::Confirmed, confirmed_messages.into()));
 
@@ -1117,7 +670,7 @@ impl GrpcService {
                         } else {
                             let mut confirmed_messages = vec![];
                             let mut finalized_messages = vec![];
-                            if matches!(message.as_ref(), Message::Block(_)) {
+                            if matches!(&message, Message::Block(_)) {
                                 if let Some(slot_messages) = messages.get(&message.get_slot()) {
                                     if let Some(confirmed_at) = slot_messages.confirmed_at {
                                         confirmed_messages.extend(
@@ -1177,7 +730,7 @@ impl GrpcService {
         stream_tx: mpsc::Sender<TonicResult<SubscribeUpdate>>,
         mut client_rx: mpsc::UnboundedReceiver<Option<Filter>>,
         mut snapshot_rx: Option<crossbeam_channel::Receiver<Box<Message>>>,
-        mut messages_rx: broadcast::Receiver<(CommitmentLevel, Arc<Vec<Arc<Message>>>)>,
+        mut messages_rx: broadcast::Receiver<(CommitmentLevel, Arc<Vec<Message>>)>,
         debug_client_tx: Option<mpsc::UnboundedSender<DebugClientMessage>>,
         drop_client: impl FnOnce(),
     ) {
@@ -1299,7 +852,7 @@ impl GrpcService {
 
                         if commitment == CommitmentLevel::Processed && debug_client_tx.is_some() {
                             for message in messages.iter() {
-                                if let Message::Slot(slot_message) = message.as_ref() {
+                                if let Message::Slot(slot_message) = &message {
                                     DebugClientMessage::maybe_send(&debug_client_tx, || DebugClientMessage::UpdateSlot { id, slot: slot_message.slot });
                                 }
                             }